+++ ChangeLog.md
@@ -6,8 +6,25 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 <details>
 <summary>1.x Releases</summary>
 
-### 1.4.0 - 2025-08-10
+### 1.4.1 - 2025-09-20
+
 #### Changed
+
+- Updated **Avalonia** to `11.3.6`.
+- Updated **NLog** to `6.0.4`.
+- Updated **MangaAndLightNovelWebScrape** to `5.0.2`.
+- Removed `Indigo` website support
+- Removed `Waterstones` website support
+- Removed `FireFox` browser support
+
+#### Fixed
+
+- Removed `MangaAndLightNovelWebScrape` logging
+
+### 1.4.0 - 2025-08-10
+
+#### Changed
+
 - Updated **Avalonia** to `11.3.3`.
 - Updated **NLog** to `6.0.3`.
 - Added search when adding a new series: a list appears on the left; select an item to use add it.
@@ -15,6 +32,7 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - During price analysis, other controls are disabled until the scrape completes.
 
 #### Fixed
+
 - ComboBox & ListBox selection styling.
 - `Hentai` filter is now selectable.
 - Crash when selecting a genre filter (introduced in the last version).
@@ -24,15 +42,20 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - Price analysis window resets correctly if the scrape fails.
 
 ### 1.3.1 – 2025-06-27
+
 #### Changed
+
 - Updated all **Avlonia** libs to `11.3.2`
 - Updated logging format message
 
 #### Fixed
+
 - Changing rating or volumes read for a series no longer clears/resets the value amount
 
 ### 1.3.0 – 2025-06-22
+
 #### Changed
+
 - Updated **NLog** to `6.0.0`
 - Added **Hentai** to genre list
 - Added **Supervisor** as a valid staff role
@@ -46,11 +69,13 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - Added supports for importing collections from Libib via their csv export
 
 #### Performance Improvements
+
 - Improved performance of the AniList client
 - Improved performance of the MangaDex client
 - Improved cover-image downloading
 
 #### Fixed
+
 - Series with missing cover images now automatically retrieve a new cover on refresh
 - Clicking a button for an already open window now sets the focus to that window
 - Fixed currency-symbol placement in collection and series value amounts based on locale
@@ -61,7 +86,9 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - Series description now defaults to the original language if an English description is not available for MangaDex entries
 
 ### 1.2.0 – 2025-06-14
+
 #### Changed
+
 - Updated **System.Drawing.Common** to `v9.0.6`  
 - Updated **System.Linq.Dynamic.Core** to `v1.6.6`  
 - Updated **Microsoft.Extensions.Http** to `v9.0.6`  
@@ -69,12 +96,15 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - Updated **Publisher, Staff, & Description** series card text to be bold
 
 #### Fixed
+
 - Resolved issue where refreshing a series with a changed status would break the UI layout until hovered  
 - Fixed bug where changing a cover via URL failed due to folder permission errors  
 - Fixed issue where a successful refresh would delete the cover filename  
 
 ### 1.1.0 – 2025-06-08
+
 #### Changed
+
 - Updated Avalonia to **v11.3.1**  
 - Updated DiscordRichPresence to **v1.3.0.28**  
 - Updated DynamicData to **v9.4.1**  
@@ -87,6 +117,7 @@ This project adheres to [Semantic Versioning](https://semver.org/) and follows t
 - Renamed **Mean Score** to **Mean Rating** in the Statistics window.
 
 #### Fixed
+
 - Make series title text copyable by clicking on it.  
 - Correct Mean Rating calculation logic.  
 - Ensure the log file is created on first run.